pyproject.toml
[tool.poetry]
name = "treadmill"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
csafe = {git = "https://gitlab.com/teriyakisaurus/python-csafe.git", rev = "main"}
nicegui = "^1.3.9"
pyserial = "^3.5"
rpi-gpio = "^0.7.1"
pygame = "^2.5.1"
evdev = "^1.6.1"
python-box = "^7.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
